Move constants loading logic to ConstantsReader, update directories
